+++ spark/src/main/scala/geotrellis/spark/store/file/cog/FileCOGLayerWriter.scala
@@ -79,16 +79,6 @@ class FileCOGLayerWriter(
         val path = s"${keyPath(key)}.${Extension}"
 
         mergeFunc match {
-          case None =>
-            cog.write(path, true)
-            // collect VRT metadata
-            (0 until cog.bandCount)
-              .map { b =>
-                val idx = Index.encode(keyIndex.toIndex(key), maxWidth)
-                (idx.toLong, vrt.simpleSource(s"$idx.$Extension", b + 1, cog.cols, cog.rows, cog.extent))
-              }
-              .foreach(samplesAccumulator.add)
-
           case Some(_) if !uriExists(path) =>
             cog.write(path, true)
             // collect VRT metadata
@@ -108,6 +98,16 @@ class FileCOGLayerWriter(
               .map { b =>
                 val idx = Index.encode(keyIndex.toIndex(key), maxWidth)
                 (idx.toLong, vrt.simpleSource(s"$idx.$Extension", b + 1, merged.cols, merged.rows, merged.extent))
+              }
+              .foreach(samplesAccumulator.add)
+
+          case _ =>
+            cog.write(path, true)
+            // collect VRT metadata
+            (0 until cog.bandCount)
+              .map { b =>
+                val idx = Index.encode(keyIndex.toIndex(key), maxWidth)
+                (idx.toLong, vrt.simpleSource(s"$idx.$Extension", b + 1, cog.cols, cog.rows, cog.extent))
               }
               .foreach(samplesAccumulator.add)
         }